Move routes to interface layer
